+++ CHANGELOG.md
@@ -10,9 +10,15 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### CHANGED
 
+- There are now both `Sync` and `!Sync` versions of the pool. The latter, in
+  `refpool::unsync::Pool`, is as performant as previously, while the thread safe
+  version in `refpool::sync::Pool` is roughly 5-10x slower, but still manages to
+  be about 25% faster than the Windows system allocator. You should prefer not
+  to use it on platforms with faster system allocators, such as Linux. To enable
+  the thread safe version, use the `sync` feature flag.
 - The pool is now reference counted, which means your `PoolRef`s won't suddenly
   become dangerously invalid when the pool goes out of scope. This also means
-  that when you clone a `Pool`, you get another reference to the same pool.
+  that you can now clone a `Pool` and get another reference to the same pool.
 
 ## [0.1.0] - 2019-11-26